Adding updateschema.pl and updateschema.data.  Thus begins an effort to automate schema changes so that mere mortals do not have to enter commands manually.
updateschema.pl is just a script that parses updateschema.data and 
communicates with the DB.  updateschema.data holds all the data regarding
schema changes.  It allows you to do run SQL and shell commands.  The 
syntax is roughly:

  * Anything beginning with '#' is a comment.
  * Anything beginning with '#%' is a command to run via shell.  I will
    probably change this to '$'.
  * Everything else is assumed to be a valid SQL command and is executed
    via DBI.

I've run it a few times on my workstation and it appears to work.  More 
testing needed.