Workaround for stupid httpd on solaris.  Consider:
           <select name="_person">
               <option value="0">[null]</option>
           </select>
           <input name="_Person" value="New" type="submit">

When you click on New, the arg value pair is _person = New, not _Person = New. 
That is a problem.  This was corrected before by prefixing table names
with '_' for New buttons.  However, the javascript block has added a variable
with the same name, but different case.  So...I'm working around this again.