This change includes a number of things.  First of all, it is now actually searching tables looking for information.  It searches the following:
  columns name in Connection
  columns name, aliases in Device
  columns name, aliases in Entity
  columns lastname, firstname, aliases in Person
  columns name, aliases in Site

Obviously, this is a subset of information.  It then looks up the 
entity associated with each hit.  If just one entity is found, that will
be displayed (code for that not in place yet).  If more than one is found,
it presents a list using sortresults.mhtml for the user to select.  

Some consideration should be devoted to what columns to search in what tables.
At one end of the extreme is searching all columns in the above tables
(more specifically, tables with a link to Entity).  At the other end is the
approach above.  It may be interesting to search the columns defined to 
be unique and those that are indexed.  Related to this is whether this
information is easily discoverable via the API.

Other tables that could be searched include BGPPeering and Ipblock, thought I 
don't think they would be useful for searching.

Lastly, I still need to tweak the layout.  I'm borrowing heavily from 
device.html, but something doesn't look right.  This may be something to 
punt on until we can a good css implementation.  Controlling layout with
tables sucks.  :-P