Lot more work on the header section so that it degrades nicely as the window is resized to something smaller.  The tabs could use some help in this regard.
Other changes were to the look of the content area and the tabs.  I'm
leaving the size of the text in the tabs as is -- they're not just linkage
but also the title of the page, obviating the need for repeating the
title of the page just under the tabs.

Does anyone want to start breaking stuff and push css into the actual pages?