Crapload of changes that are a bit too much to enumerate.  :-(  The diff on this won't be pleasant.
General idea is that we needed to break away from how the tabs were done
previously as it wouldn't work well with nested tabs.  The nested tabs
gets us what we want where each page has sub functions.

I broke what worked, but would not work in the future, and replaced it
with something that is broken, but will eventually work.

Any questions?  :-P