major CSS changes to fix the following:   fixed multiple pages from having content or error messages shifted to the right   fixed the footer from appearing in the middle of the page   now editing forms don't explode outside of their div container
In order to fix these problems, I got rid of the large containing box which was supposed to make a background and border around the entire page. This in turn implied that the header needed to be changed. I also changed the colors around because I didn't think the tan background worked without the blue grey behind it. I didn't intend them to be the UO colors, they just turned out that way, I swear! :)