In response to ticket #82:
Fixed the funny raised title bar by changing the look of it slightly.

As for the fieldsets being all on separate lines, I have no idea why Opera 7.54 does this, but Opera 8 displays it correctly. It might be a bug in the CSS code, or it might be a bug in the browser. Either way, it shouldn't be too much trouble for the few Opera users to upgrade to 8.