Searching is for the most part cleaned up.  I still need to work on the actual display of content (the what and how).
One (potential) problem I have found is the (unexpected) behavior of
sortresults.mhtml.  If the table is a Device or Ipblock,
sortresults.mhtml ignores the page option and redirects to device.html
and ip-ws.html.  I'm not sure this is the desired behavior.  My concern
is that you may be breaking the user experience.  If I'm searching for
operational information (whatever that may be) regarding uonet8-gw, I
want the most relevant bits.  If I need additional detail, I can go to
the management page.  If however you redirect me to that page, I
now have to search through fields that are not immediately necessary to
the task at hand.

The use-case for operations.html is that of a device or interface going
down.  What is that device?  Where is it?  Who has physical access to
the room or building?  Who are the points of contact?  What is that
device connected to and does it have OOB access?  There are variations
on this use-case, but the goal is to have that information displayed in
one step.