Removed all tables except one used to impose a two column layout.
This was by far the most confusing file I've worked on.  There was an
if statement with more than 300 lines of contents and in one place
there are more than 100 lines of code devoted to calculating a single
entry in a table.

IIRC one of the reasons we are moving to css based layout is because
it is more maintainable (don't have to worry about all those </?t.>
tags.  This is probably obvious, but I think much more could be gained
(in the short run) by factoring the code into many smaller
method/component calls.  Aaron has already done this with the table
generation which is a good start.  There are probably a lot of other
ways, maybe less obvious, that similar bits of code are being written
explicitly over and over again which would be much better factored out
to calls.

So what is the point of my rambling?  I think refactoring the code
should be considered at the same or a higher priority than eliminating
all the tables (in the process the tables would probably disappear
anyway).  Just my $0.2.