Closet now only has two tables, both for two column layout.
The backbone and horizontal cable tables were not shown if there was
nothing to put in them, but the boxes have "new" buttons, and so
presumably someone might like to add one if there weren't any so I
made the buttons always appear.  Let me know if this was a stupid
idea.


Fixed a bug in cable_horizontal which I caused when I was removing
tables.