Removed all embedded tables.  There were a bunch of helpful messages explaining what the fields were for which I also removed as none of the other forms have such messages and so my guess was that this is a very old interface.  Let me know if that was a bad idea.
Also fixed a minor bug where an error would occur when a devices ip
address was looked up but didn't exist.  I just added a check that it
was defined and so the fancy error handling jumps in if it fails (vs
the stack trace).  To see the original error go to netdot-test and try
to click the [update] link for the HRIS-private device (on the
management->devices page).