Login that lets the user continue what they were doing.  Uses the apache request objects knowledge of previous requests to reconstruct the request which authcookie intercepted.
Added some radio buttons so the user could decide to just go to the
main page if they want.  Don't think they would, but if this is broken
for some reason it will allow them to do what they wanted to do.

Probably a good idea to make the timeout (NetdotExpires variable in
httpd.conf) short while this is being tested so that it will happen
often enough that any problems will be revealed.