Arbitrary splitting now happens quickly as long as there isn't a huge amount of iteration that needs to be done (this case was slow before). The only way to really force a huge amount of iteration is to break out a large chunk of a container and then ask for a small chunk with the same strategy, since in this case all the possible small chunks that fall in the large chunk and get disqualified are tried.
Sometimes it seems like the something is incorrectly disqualifying
addresses (I see a "Block allocations only allowed under Container
blocks" in the logs).  Shouldn't have any thing to due with the new
code though as this is an issue with the parents type and the new code
doesn't effect the parents type.