Made the '/' in the netdot path the responsibility of the NetdotPath variable and removed the '/' in the code which uses NetdotPath in header.mhtml and login.html.
Strangely if the you made the NetdotPath ' ' (mod_perl won't let you
do a blank PerlSetVar) it would work sometimes, but sometimes the ' '
would get turned into a '+'.  Double slashes ('//'), which resulted if
you tried to set NetdotPath to '/' before, wouldn't work, but it isn't
clear why.  If you type a request with extra slashes into your browser
address bar there is no problem, and checking the apache access log
indicates they are preserved in the request.  But, if they appear in the
the html as an img or href target, the browser fails to retrieve them.
Very confusing.