Added the page htdocs/database-graph.html which displays the schema graph and has a button for regeneration.  It isn't linked to from anywhere in netdot, but regular users probably don't care about it anyway.
htdocs/database-graph.html creates its graph in htdocs/img/graphs.  It
needs write permission to do this.  I updated Makefile accordingly.

lib/GraphVizClassDBI.pm is graph generating script turned into a
package (docs and some subs, and a little generality, instead of no
subs and netdot stuff hard coded).  

lib/Makefile now knows to install lib/GraphVizClassDBI.pm

Moved a misplaced =cut in lib/Netdot/CablePlantManager.pm.  Has
nothing to do with this other stuff, but is so minor that it would be
weird to give it its own revision.  Will do it that way in the future
though, if that is preferred.