i) Made a macro called substitute which does the PREFIX, use lib .*, and the like substitutions.  The macro is in a new makefile called utility-Makefile in etc/.  The replacement was hardcoded into 6 makefiles before, now they use the macro.  For this to work they need to import the utility-Makefile.  To know where to import it from I added a ROOT variable to the toplevel makefile which is set to the cwd when the the top level makefile is run.  There is probably a built in way to do that, but I didn't come across it.  If anyone knows how, consider changing it.
ii) The macro replaces occurances of <<Make:NAME>> with $(NAME), where
NAME represents an arbitrary string, if NAME is one of the strings in
the list the REPLACABLES defined in utility-Makefile.  So, if you want
a new name to be replaced, just add it to the list in
utility-Makefile.  Because many files contained PREFIX and friends, I
had to replace them with <<Make:PREFIX>>.  That is why so many none
makefiles were changed.