Merging fix from trunk.
Fix for ticket #270.  Now all children of the parent block are skipped
over while searching for a free block in auto_allocate.  This also
fixes the speed issue which existed before (e.g., allocating a /17 and
then a /31 using the same strategy used to result in auto_allocate
trying all 2**(31-17) /31s which the /17 conflicted with.  This could
really get out of hand with IPv6) since the time the search takes is
now bounded by the number of preexisting children, not the number of
subnets of the requesting size which would fit in the parent.

M    netdot-0.7/lib/Netdot/IPManager.pm