Noticed some weirdness when doing a search for devices.  It turns out I was hitting "Enter" when I wanted to search for devices.  However, hitting enter submits the form, which resulted in the cryptic error message "Permissions are null" it took me awhile to even notice I was hitting enter sometimes and clicking others, so I tried to change the default behavior of enter when the search button is displayed, but this turned out to be pretty difficult.  So, for now, I have changed the error messages depending on what the user forgot to submit.
Also, there was a bug with checking if @object_id was defined, for some reason this would return true even if there 
were no entries in the list.  I have fixed it so that the program checks to see if the first element is defined.

Related to @object_id bug, sometimes submitting with no objects selected will grant the user permissions on volcanology-hub1.uoregon.edu
I have not been able to reporduce this since improving the error correction at the begenning of the submit section.