adding support for more mac-address-table formats
Had to make the regex pretty generic because of Cisco changing it on
many platforms.  However, the order of MAC dynamic...interface<EOL>
seems to be universal from what I've found, and there doesn't seem to be
false positives, so I'm hoping it should be a safe change.

I've also added unit tests for the regexs in t/arp.t