pass a more descriptive error message for PhysAddr validation
When a routine calls PhysAddr->validate to validate an L2 address the
descriptive message is lost by just returning a boolean success or
failure. Instead allow passing in a reference to a scalar to give the
calling function an opportunity to give the user more information about
the specifics behind Netdot rejecting the MAC address they entered.

Small amounts of whitespace formatting were performed in this commit as
well as the functional changes. Changes such as removing trailing
whitespace in functions that were touched and also replacing tabs with
four spaces.