center ip address link in ipblock subnet view
When viewing a subnet via the ipblock view, it would be helpful to center the
text of the address in the table cell. Also, it would be great to make the
whole cell clickable instead of just the text. This will help those of
us who use tiny fonts to more easily click the ipblock/address.