if any of the IP objects are not addresses, then use the 'block' view
Don't just evaluate the first object of the array - this leads to issues
where a container has static addresses at the top of the addressing
block (first entry of the array) and the whole block is rendered as though
they were hosts (addresses).

Instead, default to 'address' and see if any of the objects are not
addresses. If so, then set the type to 'block' and break the loop.