factor out code to determine DNS address record label
There is a second location in management/ip.html where we need to set
the label for the DNS address record. This second location in the code
get executed when a new address is created (and there is no "id" passed
in via the arguments of the Mason component.)

The need for a second place to contain logic warrants factoring out the
logic to determine the (DNS address record) label for a given IP block
object into the class.

Add new subroutine to Netdot::Model::Ipblock, dns_address_record_label
that returns 'A' or 'AAAA' depending on the version of the object.

Also add in some unit tests.